Add support for hide_content_if_missing flag when parsing XML content
With this flag we can provide support for users wanting to hide content that
contains a prefix if a variable is missing.

Example:
If we have the following content:
{content: [{text: foo}, {var: bar, hide_content_if_missing: true}]}

and the bar variable is missing, then the prefix wouldn't be created neither.
This assumes that the hideEmptyContent-flag is set to true on the element.