Create StorageBackend system with secure and archival storage
The goal with the storage backends is to make it easier to use the same
code independently of which backend is being read from or written to.
All backends can be used as secure or archival storage. Secure storage
stores content in containers while archival storage stores data as
regular files and/or directories.