Make revoked tasks retryable
Celery does not allow revoked tasks to be retried
by keeping a list of the ids of all revoked tasks.

To get around this we use a new celery_id field on
ProcessTask which is instead used to communicate with
celery. When a task is retried the celery_id is updated
so that celery will not find the id in the list of
revoked tasks.