Fix XML comparison in package representations
The main change here is the addition of the CompareRepresentationXMLFiles task.

This task finds each representation mets referenced by the root mets and
compares each with their respective premis file.

To better represent the result of the validation we also now have a
"representation" column in the validation list in the task detail modal which
shows which representation that each file is in.

In the future this could be improved further by creating a task for each
representation at "task runtime" for faster retries.