Add 'Aggregating' state (#2315)
This adds a new state 'Aggregating', used by the University of
Gothenburg for AIPs that are being created from multiple SIPs
over time. IPs in this state should be in the 'Checkpoint' view
for now.

This also makes links to such packages in the archival description
lead to the view where the package can be found.

Co-authored-by: Anton Neld <anton.neld@gu.se>